refactor(track): hoist normalCap to TrackParams, split domain file
Address the two-agent review of the #198 guard fix:

- normalCap (auto_tracks ? .max : count) was re-derived at both
  the layout math and the swap command, the exact drift the
  shared-authority extraction set out to remove. Hoist it onto
  TrackParams beside trackCap so both sites read one rule.
- TrackLayout+Domain.swift crossed the 350-line hard ceiling
  after the guard extraction. Split the mutating `extension
  Space` track-state verbs into TrackLayout+SpaceState.swift;
  the domain file stays the actor-free rule authority.

No behavior change.
